fix(F-123): refactor the service module
