+++ backend/app/services/scenario_preparation/__init__.py
@@ -1,0 +1,3 @@
+from app.services.scenario_preparation import scenario_preparation_service
+
+__all__ = ['scenario_preparation_service']
