+++ backend/app/services/session_feedback/__init__.py
@@ -1,0 +1,3 @@
+from app.services.session_feedback import session_feedback_service
+
+__all__ = ['session_feedback_service']
